refactor endpoint network request send
